Show error state when searching for gtin/sku fails
